Prevent init of default when not necessary Move test
