+++ src/AutoMapper/TypeMap.cs
@@ -76,7 +76,14 @@ public class TypeMap
     public MemberList ConfiguredMemberList
     {
         get => (_details?.ConfiguredMemberList).GetValueOrDefault();
-        set { Details.ConfiguredMemberList = value; }
+        set 
+        {
+            if (_details == null && value == default)
+            {
+                return;
+            }
+            Details.ConfiguredMemberList = value;
+        }
     }
     public IReadOnlyCollection<TypePair> IncludedDerivedTypes => (_details?.IncludedDerivedTypes).NullCheck();
     public IReadOnlyCollection<TypePair> IncludedBaseTypes => (_details?.IncludedBaseTypes).NullCheck();
